+++ Runtime/Core/Ease.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace FlowTween {
+    
+[Serializable]
+public struct Ease {
+    public EaseDirection Direction;
+    public EaseType Type;
+}
+
+}